refactor: change folder in pages
